Update Readme and build scripts

+++ microros_static_library_ide/library_generation/library_generation.sh
@@ -1,16 +1,22 @@
 #!/bin/bash
 set -e
 
-export BASE_PATH=/project/$MICROROS_LIBRARY_FOLDER
+export BASE_PATH=$(pwd)/..
+export PROJECT_PATH=../../../
 
 ######## Check existing library ########
-if [ -f "$BASE_PATH/libmicroros/libmicroros.a" ]; then
+if [[ -f "$BASE_PATH/libmicroros/libmicroros.a" && ! -f "$USER_COLCON_META" ]]; then
     echo "micro-ROS library found. Skipping..."
-    echo "Delete $MICROROS_LIBRARY_FOLDER/libmicroros/ for rebuild."
+    echo "Delete microros_static_library_ide/libmicroros/ for rebuild."
     exit 0
 fi
+
+if [ ! -f "$USER_COLCON_META" ]; then
+    rm -rf $BASE_PATH/libmicroros
+fi
+
 ######## Trying to retrieve CFLAGS ########
-export RET_CFLAGS=$(find /project -type f -name *.mk -exec cat {} \; | python3 $BASE_PATH/library_generation/extract_flags.py)
+export RET_CFLAGS=$(find $PROJECT_PATH -type f -name subdir.mk -exec cat {} \; | python3 $BASE_PATH/library_generation/extract_flags.py)
 RET_CODE=$?
 
 if [ $RET_CODE = "0" ]; then
@@ -23,66 +29,38 @@ else
     exit 1;
 fi
 
-######## Init ########
-apt-get update
-apt-get install -y gcc-arm-none-eabi
-
-cd /uros_ws
-
-source /opt/ros/$ROS_DISTRO/setup.bash
-source install/local_setup.bash
-
-ros2 run micro_ros_setup create_firmware_ws.sh generate_lib
-
-######## Adding extra packages ########
-pushd firmware/mcu_ws > /dev/null
-
-    # Workaround: Copy just tf2_msgs
-    git clone -b ros2 https://github.com/ros2/geometry2
-    cp -R geometry2/tf2_msgs ros2/tf2_msgs
-    rm -rf geometry2
-
-    # Import user defined packages
-    mkdir extra_packages
-    pushd extra_packages > /dev/null
-        USER_CUSTOM_PACKAGES_DIR=$BASE_PATH/../../microros_component/extra_packages 
-    	if [ -d "$USER_CUSTOM_PACKAGES_DIR" ]; then
-    		cp -R $USER_CUSTOM_PACKAGES_DIR/* .
-		fi
-        if [ -f $USER_CUSTOM_PACKAGES_DIR/extra_packages.repos ]; then
-        	vcs import --input $USER_CUSTOM_PACKAGES_DIR/extra_packages.repos
-        fi
-        cp -R $BASE_PATH/library_generation/extra_packages/* .
-        vcs import --input extra_packages.repos
-    popd > /dev/null
-
-    if [ ! -z ${MICROROS_USE_EMBEDDEDRTPS+x} ]; then
-        rm -rf eProsima/Micro-XRCE-DDS-Client
-        rm -rf uros/rmw_microxrcedds
-
-        git clone -b main https://github.com/micro-ROS/rmw_embeddedrtps uros/rmw_embeddedrtps
-        git clone -b main https://github.com/micro-ROS/embeddedRTPS uros/embeddedRTPS
-    fi
-
-popd > /dev/null
+echo "Using:"
+echo "-------------"
+echo $(which arm-none-eabi-gcc)
+echo Version: $(arm-none-eabi-gcc -dumpversion)
+echo "-------------"
 
 ######## Build  ########
-export TOOLCHAIN_PREFIX=/usr/bin/arm-none-eabi-
 
-if [ ! -z ${MICROROS_USE_EMBEDDEDRTPS+x} ]; then
-    ros2 run micro_ros_setup build_firmware.sh $BASE_PATH/library_generation/toolchain.cmake $BASE_PATH/library_generation/colcon-embeddedrtps.meta
+if [ ! -f "$BASE_PATH/libmicroros/libmicroros.a" ]; then
+    # If library does not exist build it
+
+    ######## Add extra packages  ########
+    pushd extra_packages
+        # Workaround: Copy just tf2_msgs
+        git clone -b ros2 https://github.com/ros2/geometry2
+        cp -R geometry2/tf2_msgs tf2_msgs
+        rm -rf geometry2
+
+        if [ -f $extra_packages.repos ]; then
+        	vcs import --input extra_packages.repos
+        fi
+    popd > /dev/null
+
+
+    make -f libmicroros.mk
 else
-    ros2 run micro_ros_setup build_firmware.sh $BASE_PATH/library_generation/toolchain.cmake $BASE_PATH/library_generation/colcon.meta
+    # If exists just rebuild
+    make -f libmicroros.mk rebuild_metas
 fi
 
-find firmware/build/include/ -name "*.c"  -delete
-rm -rf $BASE_PATH/libmicroros
-mkdir -p $BASE_PATH/libmicroros/include
-cp -R firmware/build/include/* $BASE_PATH/libmicroros/include/
-cp -R firmware/build/libmicroros.a $BASE_PATH/libmicroros/libmicroros.a
-
 ######## Fix include paths  ########
-pushd firmware/mcu_ws > /dev/null
+pushd $BASE_PATH/libmicroros/micro_ros_src > /dev/null
     INCLUDE_ROS2_PACKAGES=$(colcon list | awk '{print $1}' | awk -v d=" " '{s=(NR==1?s:s d)$0}END{print s}')
 popd > /dev/null
 
@@ -92,16 +70,3 @@ for var in ${INCLUDE_ROS2_PACKAGES}; do
         rm -rf $BASE_PATH/libmicroros/include/${var}/${var}
     fi
 done
-
-######## Generate extra files ########
-find firmware/mcu_ws/ros2 \( -name "*.srv" -o -name "*.msg" -o -name "*.action" \) | awk -F"/" '{print $(NF-2)"/"$NF}' > $BASE_PATH/libmicroros/available_ros2_types
-find firmware/mcu_ws/extra_packages \( -name "*.srv" -o -name "*.msg" -o -name "*.action" \) | awk -F"/" '{print $(NF-2)"/"$NF}' >> $BASE_PATH/libmicroros/available_ros2_types
-
-cd firmware
-echo "" > $BASE_PATH/libmicroros/built_packages
-for f in $(find $(pwd) -name .git -type d); do pushd $f > /dev/null; echo $(git config --get remote.origin.url) $(git rev-parse HEAD) >> $BASE_PATH/libmicroros/built_packages; popd > /dev/null; done;
-
-######## Fix permissions ########
-sudo chmod -R 777 $BASE_PATH/libmicroros/
-sudo chmod -R 777 $BASE_PATH/libmicroros/include/
-sudo chmod -R 777 $BASE_PATH/libmicroros/libmicroros.a
